==> Notebooks/Thomas/1 - Intro to yfinance.ipynb <==
# Introduction to google Colab and yFinance.

Good morning all, I am just putting together this quick notebook so that we can reference it if we need to. This will allow you to get comfortable with the yfinance API and it will also allow us to figure out how to connect to the shared network with our notebooks. The fact that the Machine Learning Pipelines is one of the prereq's I imagine we should use the markfile setup with multiple python scripts that can be run alone or run together with the if __name__ == "__main__": command. So let's get started with connecting.

# Just going to put an easy dynamic variable to change with whoever is working in the notebook

user = 'Tom'

## Connecting Colab to Shared Drive
We will start by adding a few libraries and then setting the file path. The file path will likely be different for all of us so I will make it dynamic and you can put in the pathway you have and we can just switch out the names.

# importing only the connection libraries
import os
from google.colab import drive
drive.mount('/content/drive')

#Specify paths specific to the user
if user.lower() == 'tom':
    path = '/content/drive/MyDrive/Colab Notebooks/696 - Milestone II/696 - Milestone II - Shared'
if user.lower() == 'peter':
    path = ''
if user.lower() == 'melody':
    path = ''

#Change to the directory for the shared network
os.chdir(path)
os.getcwd()

#Let's make sure its connected to the right one by listing some files in the directory
os.listdir()

## Introduction to yfinance
yfinance is relatively easy to use and one of the most popular free sources of stock data. I used it in milestone I to obtain price data for thousands of stocks (be careful of rate limiting) and I am excited to explore more of the fundamental side of the API with this project. I'll show you a very quick and easy example but feel free to check out the documentation [here](https://https://ranaroussi.github.io/yfinance/).


# Import the libraries for working with yfinance
import numpy as np
import pandas as pd
import yfinance as yf
import altair as alt

#Let's start with a really popular ticker for Microsoft
'''Just as a side note, a ticker is the stock symbol, which is typically a three to four letter code to identify the stock on the stock exchange. We will be dealing with them a lot just so you get the terminology'''

#Assign our ticker
msft = yf.Ticker('MSFT')

#Let's start by getting some price data. Let's do the past three weeks in 5' intervals (it will give us the date/time, high, low, open, close and volume for every 5 mins of trading over the past three weeks)
msft_price = yf.download(tickers = 'MSFT', period = '21d', interval = '5m')
msft_price.head()

Great, so it was that simple to get historical price data for a stock. Now, let's get some fundamental information like the sector, market cap, etc. These are things that we will likely want to add to the dataframe for our machine learning pipeline. so it's pretty easy to pull the info we want. Let's just look at a few.

#Get the corporate info for Microsoft
msft_info = msft.info
#Let's print off some easy information
print(f'Sector: {msft_info.get('sector')}')
print(f'Industry: {msft_info.get('industry')}')
print(f'Market Capitalization: {msft_info.get('marketCap')}')
print(f'Employees: {msft_info.get('fullTimeEmployees')}')

Now, for the big part of our work. We are going to be working a lot with financail statements in order to model revenue, expenses, margin, and other health indicators for a company moving forward so that we can predict a price target for the share price in the future. So let's grab some annual statements and some quarterly statements.

#Let's pull some financial information from Microsoft
annual_statements = msft.get_income_stmt(freq='yearly')
quarterly_statements = msft.get_income_stmt(freq='quarterly')

Now, let's take a look at both of these dataframes. We will need to figure out which metrics are good, which are likely unimportant (or highly correlated with other metrics). We can then also think about normalization (percentage of revenue) so that companies are comparable regardless of market cap and a way of bring in timeseries analysis (using year over year change instead of just the raw values alone). These are just my thoughts so far.

annual_statements

Let's look at the revenue for the past four years and see if it is growing and where we would forecast it to go just by eye.

plotting_df = annual_statements.T
plotting_df.reset_index(inplace=True, names = 'date')
#plotting_df.head()
rev = alt.Chart(plotting_df).mark_bar(width=20, color = 'black').encode(x='date:T', y='TotalRevenue:Q')
rev.show()

quarterly_statements

Okay, I think that is a good introduction and hopefully it gives you something to play around with and look at. Let me know if you have any questions!
