==> Notebooks/Thomas/3 - Macro_Data_Fred.ipynb <==
# Fred API to get the data we need

ALright, so far we have only been point and clicking in order to download the financial data and then we merge it to the dataset. I want to use the API to get the data so it is autoated in our pipeline. I am going to use this notebook to get the right code to build it.

Info We need:

Quarters:  
_2024Q2  
_2024Q3  
_2024Q4  
_2025Q1  
_2025Q2  

Data:  
GDP  
Unemployment  
InterestRate  
IndustrialProd  
Inflation  

Alright, I have gotten an API from FRED, so now we can try and build a single row with multiple columns. It's probably easiest to build a dataframe and then just get the transpose. Let's get started.


#pip install fredapi

#Libraries
import pandas as pd
from fredapi import Fred


#API Key (Free to obtain)
fred_api_key = "0fd09a8eca4539118cb3699daec8667a"

Alright, now let's try to get all of the data that we need.

fred = Fred(api_key = fred_api_key)
GDP = fred.get_series("GDP")[-6:]
GDPReal = fred.get_series("GDPC1")[-6:]
Unemployment = fred.get_series("UNRATE").resample("QS").mean()[-7:-1]
InterestRate = fred.get_series("FEDFUNDS").resample("QS").mean()[-7:-1]
IndustrialProd = fred.get_series("INDPRO").resample("QS").mean()[-7:-1]
Inflation = fred.get_series("CPIAUCSL").resample('QS').mean()[-7:-1]
print(f"Last 6 GDP:\n{GDP}")
print(f"Last 6 Industrial Prod:\n{IndustrialProd}")

Alright, well that was easier than I expected. Now we just have to create the dataframe and add all of the info and then get the transpose.

row_data = {
    "GDP": GDP,
    "GDPReal": GDPReal,
    "Unemployment": Unemployment,
    "InterestRate": InterestRate,
    "IndustrialProd": IndustrialProd,
    "Inflation": Inflation,
}
quarters = ['_2024Q1','_2024Q2','_2024Q3','_2024Q4','_2025Q1']

macro = {f"{name}{q}": float(v)
         for name, s in row_data.items()
         for q, v in zip(quarters, s)}

macro_data = pd.DataFrame([macro])
macro_data


Alright, well that wasn't as difficult as I expected it to be. So we can move on to building it into our machine learning pipeline.
