==> Notebooks/Thomas/5 -Train Test Split Notebook.ipynb <==
# Train - Test - Split
We are going to split the data into Training and Testing CSV's early to make sure that we have no data leakage. We are going to do this even before our imputation step because we don't want any of the information in the Test dataset being used to affect the training data set. This is going to be a quick notebook to run through the code necessary to create the train.csv and the test.csv files that we can further work with.

# File system libraries
import os
from google.colab import drive

# Data Manipulation Libraries
import numpy as np
import pandas as pd

# Stat Libraries
import scipy.stats as stats

# Machine Learning Libraries
#import pycaret #Not working with this version of python
from sklearn.model_selection import train_test_split

# Visualization Libraries
import matplotlib.pyplot as plt
import seaborn as sns
import altair as alt

# show decimals without scientific notation
pd.set_option('display.float_format', '{:,.2f}'.format)

# Mount the google drive
drive.mount('/content/drive')
# Navigate to the folder and set the file name
path = '/content/drive/MyDrive/Colab Notebooks/696 - Milestone II/696 - Milestone II - Shared/Pipeline Files'

os.chdir(path)
os.getcwd()
os.listdir()

filename = 'Russell_3000_Cleaned.csv'
dataset = pd.read_csv(filename)
print(dataset.shape)
print(dataset.head())

Alright, now that we have the data imported properly. We are going to split this into two files. A train file that has both the independent (X) variables and the dependent (y) variables we are trying to predict. We will have the same for the test file. We will then need to merge (X)(y) back together before saving them into their independent files. The first thing we are going to need to do is seperate the (X) columns from the (y) columns. So let's work on that to start.

y_columns = []
X_columns = []
columns = dataset.columns
for column in columns:
    if '_2025Q2' in column:
        y_columns.append(column)
    else:
        X_columns.append(column)
# Quick sanity check
for column in X_columns:
    print(column)

Alright, now that we have the columns we can start splitting the data based on both the sector and market cap. Let's start by creating the combined column.

# Create the startum
strat = (dataset['Sector'].astype(str) + ' | ' + dataset['Market Cap'].astype(str))

# We need to drop anything that is too small
val_cou = strat.value_counts()
val_cou

# Based on this I would say we need at least 4 to stratify
to_strat = strat.isin(val_cou[val_cou >= 2].index)
# Now, let's pull out those
dataset, strat = dataset.loc[to_strat],strat.loc[to_strat]

X = dataset[X_columns]
y = dataset[y_columns]

X_tr, X_te, y_tr, y_te, strat_tr, strat_te = train_test_split(X, y, strat, test_size = 0.2, random_state = 6, stratify = strat)


Okay, so that should have stratified everything but let's do a quick test to see if we have balance.

# Training Set
print(pd.crosstab(dataset['Sector'].loc[strat_tr.index],
                  dataset['Market Cap'].loc[strat_tr.index], normalize="all").round(3))
# Testing Set
print(pd.crosstab(dataset['Sector'].loc[strat_te.index],
                  dataset['Market Cap'].loc[strat_te.index], normalize="all").round(3))

That looks pretty darn good in terms of the stratification. Now we can just export them to their own CSV files.

# If we want to export to four different files, we can do that.
#X_tr.to_csv("X_train.csv", index=True)
#y_tr.to_csv("y_train.csv", index=True)
#X_te.to_csv("X_test.csv",  index=True)
#y_te.to_csv("y_test.csv",  index=True)

# Or, if we want to merge them back together before saving, we can do that too.
train = pd.concat([X_tr,y_tr],axis = 1)
test = pd.concat([X_te,y_te],axis = 1)

train.to_csv('train.csv', index = True)
test.to_csv('test.csv',index = True)

Now that we have updated the pipeline and run the script, lets take a look at the shape of all of our files to make sure they are what we expect.

X_train = pd.read_csv('X_train.csv')
y_train = pd.read_csv('y_train.csv')
X_test = pd.read_csv('X_test.csv')
y_test = pd.read_csv('y_test.csv')

print(X_train.shape)
print(y_train.shape)
print(X_test.shape)
print(y_test.shape)

Looks good to me, I don't want to investigate further because I don't want to destroy the integrity of the test split.
