==> Notebooks/Thomas/6 - Feature_Engineering_Part_II.ipynb <==
# Feature Engineering (Part II)

Update 9-25-25: I am adjusting this notebook to work with the new KPI files from the GitHub Repository.

This notebook will be used to build the function that will calculate all of the quarterly percentage changes and rate of change features. It will add the data back to the file. Consider calcualtions for percent of revenue.


# File system libraries
import os
#from google.colab import drive

# Data Manipulation Libraries
import numpy as np
import pandas as pd

# Stat Libraries
import scipy.stats as stats

# Machine Learning Libraries
#import pycaret #Not working with this version of python
import sklearn

# Visualization Libraries
import matplotlib.pyplot as plt
import seaborn as sns
import altair as alt


# show decimals without scientific notation
pd.set_option('display.float_format', '{:,.2f}'.format)

# Mount the google drive
# drive.mount('/content/drive')
# Navigate to the folder and set the file name
path = '../../datasets'

os.chdir(path)
os.getcwd()
os.listdir()

train_file = 'X_train_filled_KPIs.csv'
test_file =  'X_test_filled_KPIs.csv'
train_dataset = pd.read_csv(train_file)
test_dataset = pd.read_csv(test_file)
train_dataset.drop(columns=['Unnamed: 0'], inplace=True)
test_dataset.drop(columns=['Unnamed: 0'], inplace=True)
print(train_dataset.head())

print(train_dataset.shape)
print(test_dataset.shape)

Alright, now that we have it all loaded, we are going to have to look at quarterly changes. From the quarterly changes we can start calculating the rate of change etc. The best way to do this I think will be to design a function that will output a series that we can then just add into the dataframe. As parameters, it will take our original dataframe, the fature and the quarters to calculate the change. First, let's get a list of all the columns that we will need to calculate this for.

# Let's create a copy of the dataset so that we can merge data back to it later
original_train_df = train_dataset.copy()
columns = train_dataset.columns
for column in columns:
    print(column)

Okay, so the easy part is the quarters so let's build a list of quarters

quarters = ['_2024Q2','_2024Q3','_2024Q4','_2025Q1']

Now, we need to create a list of unique columns. We can do this by iterating trhough splitting adding to a set and then creating a list.

unique_columns = set()
for column in columns:
    column = column.replace(' ','_')
    words = column.split('_')
    if words[0] != 'KPI':
        unique_columns.add(words[0])
    else:
        unique_columns.add(str(words[0]) + '_' + str(words[1]))
unique_columns = list(unique_columns)
for column in unique_columns:
    print(column)


Great, now we can drop all of the features that are not quarterly as we won't be calculating the differences for these.

drop = {'Name','Ticker','Exchange','Sector','Market','Location','Inflation','Unemployment','IndustrialProd','GDP','GDPReal','InterestRate'}

unique_columns = [c for c in unique_columns if c not in drop]
for c in unique_columns:
    print(c)

Okay, Now we have a solid list of the raw data that we want to calcualte quarterly information for, we can start to build the function that will accomplish what we want it to. We can return a series, or we can just have the function build it into the dataset without returning anything, which is likely way more efficient. Thinking about the nesting of this, we will iterate through each unique column, we will then iterate through the quarters to calcuate the values and add it to the dataframe before moving on to the next unique column.

def quarterly_changes(dataset, unique_columns, quarters):
    for column in unique_columns:
        try:
            for val in range(1,len(quarters)):
                series_1 = dataset[str(column) + str(quarters[val-1])]
                series_2 = dataset[str(column) + str(quarters[val])]
                dataset[f'{column}_QoQ_{quarters[val-1][-4:]}_{quarters[val][-4:]}'] = (series_2 - series_1)/series_1
        except:
            for val in range(2,len(quarters)):
                series_1 = dataset[str(column) + str(quarters[val-1])]
                series_2 = dataset[str(column) + str(quarters[val])]
                dataset[f'{column}_QoQ_{quarters[val-1][-4:]}_{quarters[val][-4:]}'] = (series_2 - series_1)/series_1

    print(f"Completed Quarterly Change Calculations")
    return dataset


# Run the function to see the output
train_dataset = quarterly_changes(train_dataset, unique_columns, quarters)
print(train_dataset.shape)
train_dataset.head()

# Need to do the same with the test
test_dataset = quarterly_changes(test_dataset, unique_columns, quarters)
print(test_dataset.shape)
test_dataset.head()

Let's now take all of these rates of change, plot them and take the line of best fit so that we can get the slope as an accurate rate of change over the past 5 quarters.

# Let's get all of the columns again
columns = train_dataset.columns
# Use sets to avoid duplicates
QoQ_columns = set()
QoQ_quarters = set()
for column in columns:
    if 'QoQ' in column:
        # Pull out the common Column Name
        QoQ_columns.add(column[:-10])
        # Let's also pull out the QoQ
        QoQ_quarters.add(column[-10:])
QoQ_columns = list(QoQ_columns)
QoQ_quarters = list(QoQ_quarters)
# Order will be important so let's sort them
QoQ_quarters.sort()
QoQ_quarters


Now that we have a list of all of the features, we need to start pulling out all of the data in a funciton. This is no longer functioning because of either NaN's so we need to setup a save way to do this.

def safe_slope(vals: list[float]) -> float:
    a = np.asarray(vals, float)
    a = a[np.isfinite(a)]
    n = a.size
    if n < 2:
        return np.nan
    t = np.arange(n, dtype=float)
    
    # OLS slope without lstsq
    St, Sy = t.sum(), a.sum()
    Stt, Sty = (t*t).sum(), (t*a).sum()
    denom = n*Stt - St*St
    if np.isclose(denom, 0.0):
        return 0.0
    
    return (n*Sty - St*Sy) / denom

def get_rate_data(row, cols, quarters):
    for col in cols:
        vals = [row.get(f"{col}{q}", np.nan) for q in quarters]
        row[f"{col}_Rate"] = safe_slope(vals)
    return row

Alright now that we have the function let's apply it to our dataset.

train_dataset = train_dataset.apply(lambda r: get_rate_data(r, QoQ_columns, QoQ_quarters), axis=1)
train_dataset = train_dataset.apply(lambda r: get_rate_data(r, unique_columns, quarters), axis=1)
test_dataset = test_dataset.apply(lambda r: get_rate_data(r, QoQ_columns, QoQ_quarters), axis=1)
test_dataset = test_dataset.apply(lambda r: get_rate_data(r, unique_columns, quarters), axis=1)
train_dataset.head()

missing = train_dataset['CapitalExpenditure_QoQ_24Q4_25Q1'].isna().sum()
total = train_dataset.shape[0]
ratio =(missing/total)*100
print(f'The percent of missing data in the rates is {np.round(ratio,2)}%')

print(train_dataset.KPI_GrossProfitMargin_Rate.min())
print(train_dataset.KPI_GrossProfitMargin_Rate.max())

## Final Section to Save new CSV

# Uncomment this block when we want to save to a CSV file
#train_dataset.to_csv("X_train_filled_KPIs_QoQ.csv", index=True)
#test_dataset.to_csv("X_test_filled_KPIs_QoQ.csv", index=True)
