==> Notebooks/Thomas/7 - Troubleshooting QoQ calculations.ipynb <==
# Troubleshooting QoQ Calculations
In the cleaning file we removed all raw NaN values, so something seems to be going wrong in the QoQ feature engineering file that is leading to the introduction of novel NaN values. I'm going to use this notebook to investigate the potential reasons as to why that is happening. Let's start by importing some data-manipulation libraries.


import numpy as np
import pandas as pd

Now, let's import the files that we will be working with.

path = '../../datasets/X_train_filled_KPIs_QoQ.csv'
df = pd.read_csv(path, index_col=0)
df.head()

Let's grab all columns so that if we need to use some we can easily reference this list

columns = df.columns.tolist()
for col in sorted(columns):
    print(col)

Alright, let's now take a look at what columns have NaN values so we can start investigating why.

print(f"The column with the most NaN values has {df.isna().sum().max()} missing values.")
df.iloc[:,-80:-40].isna().sum()


Alright, so it looks like the column with the most missing values is the Gross Profit Margin Calculation. So, let's filter our dataframe using masking and then investigate the columns in this.

missing_df = df[df['KPI_GrossProfitMargin_QoQ_24Q3_24Q4'].isna()]
missing_df.shape

Let's also take a look and see if this captures most of the other NaN values in the dataframe

non_missing_df = df[~df['KPI_GrossProfitMargin_QoQ_24Q3_24Q4'].isna()]
print(f"Maximum number of missing columns after we address net margin is {non_missing_df.isna().sum().max()}")

Okay, so we will still having missing values after this that we will have to address.

columns_to_investigate = ['KPI_GrossProfitMargin_2024Q3','KPI_GrossProfitMargin_2024Q4','KPI_GrossProfitMargin_QoQ_24Q3_24Q4']
df[columns_to_investigate].sample(10)


After iterating through about 10 samples of 10, I can comfortably say that nearly all of these NaN values are due to a 0 Gross Profit Margin in the companies. So we will be able to address these pretty easily with a if denominator == 0 statement. Let's start investigating the nonmissing value dataframe for the ones that are left.

non_missing_df.iloc[:,-40:].isna().sum()

Alright, now the most values are the income tax expense QoQ rate. So we can investigate this furter the important thing here is that we are looking for all of the QoQ values in order to find the line of best fit in order to get the slope.

missing_income = non_missing_df[non_missing_df['IncomeTaxExpense_QoQ_Rate'].isna()]
missing_income.shape

columns_to_investigate = ['IncomeTaxExpense_QoQ_24Q2_24Q3','IncomeTaxExpense_QoQ_24Q3_24Q4','IncomeTaxExpense_QoQ_24Q4_25Q1','IncomeTaxExpense_QoQ_Rate']
df[columns_to_investigate].sample(20)

So here we have a bunch of NaN and inf which implies that the QoQ calculation is likely the culprit so let's investigate that further and see if it addresses this.

columns_to_investigate = ['IncomeTaxExpense_2024Q2','IncomeTaxExpense_2024Q3','IncomeTaxExpense_2024Q4','IncomeTaxExpense_2025Q1']
df[columns_to_investigate].sample(20)

Alright, once again, it looks like our big issue is when we have zeros. I think in all of our calculations we should address a 0.0 with a 0.00000001. I will adjust the QoQ functions to address this. and then run through this notebook again to see if we have addressed the issues.
